move 'minBy' tests to separate file
