+++ types/ramda/tsconfig.json
@@ -145,6 +145,7 @@
         "test/mergeRight-tests.ts",
         "test/mergeWith-tests.ts",
         "test/mergeWithKey-tests.ts",
-        "test/min-tests.ts"
+        "test/min-tests.ts",
+        "test/minBy-tests.ts"
     ]
 }
